+++ app/src/main/java/com/sopt/now/presentation/signup/SignUpActivity.kt
@@ -3,7 +3,7 @@ package com.sopt.now.presentation.signup
 import android.os.Bundle
 import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
-import com.sopt.now.core.base.AuthViewModelFactory
+import com.sopt.now.core.base.factory.AuthViewModelFactory
 import com.sopt.now.core.util.showToast
 import com.sopt.now.data.ServicePool
 import com.sopt.now.databinding.ActivitySignUpBinding
